+++ api/src/test/java/io/opencensus/metrics/MetricRegistryTest.java
@@ -20,6 +20,7 @@ import static com.google.common.truth.Truth.assertThat;
 
 import java.util.Collections;
 import java.util.List;
+import java.util.Map;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
@@ -40,12 +41,17 @@ public class MetricRegistryTest {
   private static final LabelKey LABEL_KEY = LabelKey.create("test_key", "test key description");
   private static final List<LabelKey> LABEL_KEYS = Collections.singletonList(LABEL_KEY);
   private static final LabelValue LABEL_VALUE = LabelValue.create("test_value");
+  private static final LabelValue LABEL_VALUE_2 = LabelValue.create("test_value_2");
   private static final List<LabelValue> LABEL_VALUES = Collections.singletonList(LABEL_VALUE);
+  private static final Map<LabelKey, LabelValue> CONSTANT_LABELS =
+      Collections.singletonMap(
+          LabelKey.create("test_key_1", "test key description"), LABEL_VALUE_2);
   private static final MetricOptions METRIC_OPTIONS =
       MetricOptions.builder()
           .setDescription(DESCRIPTION)
           .setUnit(UNIT)
           .setLabelKeys(LABEL_KEYS)
+          .setConstantLabels(CONSTANT_LABELS)
           .build();
   private final MetricRegistry metricRegistry =
       MetricsComponent.newNoopMetricsComponent().getMetricRegistry();