+++ impl_core/src/main/java/io/opencensus/implcore/metrics/DerivedLongGaugeImpl.java
@@ -37,6 +37,7 @@ import java.util.Collections;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Map.Entry;
 
 /*>>>
 import org.checkerframework.checker.nullness.qual.Nullable;
@@ -46,14 +47,28 @@ import org.checkerframework.checker.nullness.qual.Nullable;
 public final class DerivedLongGaugeImpl extends DerivedLongGauge implements Meter {
   private final MetricDescriptor metricDescriptor;
   private final int labelKeysSize;
+  private final List<LabelValue> constantLabelValues;
 
   private volatile Map<List<LabelValue>, PointWithFunction<?>> registeredPoints =
       Collections.<List<LabelValue>, PointWithFunction<?>>emptyMap();
 
-  DerivedLongGaugeImpl(String name, String description, String unit, List<LabelKey> labelKeys) {
-    labelKeysSize = labelKeys.size();
+  DerivedLongGaugeImpl(
+      String name,
+      String description,
+      String unit,
+      List<LabelKey> labelKeys,
+      Map<LabelKey, LabelValue> constantLabels) {
+    List<LabelValue> constantLabelValues = new ArrayList<LabelValue>();
+    List<LabelKey> allKeys = new ArrayList<>(labelKeys);
+    for (Entry<LabelKey, LabelValue> label : constantLabels.entrySet()) {
+      // Ensure constant label keys and values are in the same order.
+      allKeys.add(label.getKey());
+      constantLabelValues.add(label.getValue());
+    }
+    labelKeysSize = allKeys.size();
     this.metricDescriptor =
-        MetricDescriptor.create(name, description, unit, Type.GAUGE_INT64, labelKeys);
+        MetricDescriptor.create(name, description, unit, Type.GAUGE_INT64, allKeys);
+    this.constantLabelValues = Collections.unmodifiableList(constantLabelValues);
   }
 
   @Override
@@ -61,15 +76,17 @@ public final class DerivedLongGaugeImpl extends DerivedLongGauge implements Mete
       List<LabelValue> labelValues,
       @javax.annotation.Nullable T obj,
       ToLongFunction</*@Nullable*/ T> function) {
-    Utils.checkListElementNotNull(checkNotNull(labelValues, "labelValues"), "labelValue");
-    checkArgument(
-        labelKeysSize == labelValues.size(), "Label Keys and Label Values don't have same size.");
     checkNotNull(function, "function");
+    Utils.checkListElementNotNull(checkNotNull(labelValues, "labelValues"), "labelValue");
+    List<LabelValue> labelValuesCopy = new ArrayList<LabelValue>(labelValues);
+    labelValuesCopy.addAll(constantLabelValues);
 
-    List<LabelValue> labelValuesCopy =
-        Collections.unmodifiableList(new ArrayList<LabelValue>(labelValues));
+    checkArgument(
+        labelKeysSize == labelValuesCopy.size(),
+        "Label Keys and Label Values don't have same size.");
 
-    PointWithFunction<?> existingPoint = registeredPoints.get(labelValuesCopy);
+    PointWithFunction<?> existingPoint =
+        registeredPoints.get(Collections.unmodifiableList(labelValuesCopy));
     if (existingPoint != null) {
       throw new IllegalArgumentException(
           "A different time series with the same labels already exists.");
@@ -86,11 +103,13 @@ public final class DerivedLongGaugeImpl extends DerivedLongGauge implements Mete
 
   @Override
   public synchronized void removeTimeSeries(List<LabelValue> labelValues) {
-    checkNotNull(labelValues, "labelValues");
+    List<LabelValue> labelValuesCopy =
+        new ArrayList<LabelValue>(checkNotNull(labelValues, "labelValues"));
+    labelValuesCopy.addAll(constantLabelValues);
 
     Map<List<LabelValue>, PointWithFunction<?>> registeredPointsCopy =
         new LinkedHashMap<List<LabelValue>, PointWithFunction<?>>(registeredPoints);
-    if (registeredPointsCopy.remove(labelValues) == null) {
+    if (registeredPointsCopy.remove(labelValuesCopy) == null) {
       // The element not present, no need to update the current map of time series.
       return;
     }