+++ impl_core/src/test/java/io/opencensus/implcore/metrics/LongGaugeImplTest.java
@@ -35,6 +35,7 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
+import java.util.Map;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
@@ -57,6 +58,8 @@ public class LongGaugeImplTest {
       Collections.singletonList(LabelValue.create("value1"));
   private static final List<LabelValue> DEFAULT_LABEL_VALUES =
       Collections.singletonList(UNSET_VALUE);
+  private static final Map<LabelKey, LabelValue> EMPTY_CONSTANT_LABELS =
+      Collections.<LabelKey, LabelValue>emptyMap();
 
   private static final Timestamp TEST_TIME = Timestamp.create(1234, 123);
   private final TestClock testClock = TestClock.create(TEST_TIME);
@@ -64,7 +67,8 @@ public class LongGaugeImplTest {
       MetricDescriptor.create(
           METRIC_NAME, METRIC_DESCRIPTION, METRIC_UNIT, Type.GAUGE_INT64, LABEL_KEY);
   private final LongGaugeImpl longGaugeMetric =
-      new LongGaugeImpl(METRIC_NAME, METRIC_DESCRIPTION, METRIC_UNIT, LABEL_KEY);
+      new LongGaugeImpl(
+          METRIC_NAME, METRIC_DESCRIPTION, METRIC_UNIT, LABEL_KEY, EMPTY_CONSTANT_LABELS);
 
   @Test
   public void getOrCreateTimeSeries_WithNullLabelValues() {
@@ -80,7 +84,8 @@ public class LongGaugeImplTest {
     List<LabelValue> labelValues = Arrays.asList(LabelValue.create("value1"), null);
 
     LongGaugeImpl longGauge =
-        new LongGaugeImpl(METRIC_NAME, METRIC_DESCRIPTION, METRIC_UNIT, labelKeys);
+        new LongGaugeImpl(
+            METRIC_NAME, METRIC_DESCRIPTION, METRIC_UNIT, labelKeys, EMPTY_CONSTANT_LABELS);
     thrown.expect(NullPointerException.class);
     thrown.expectMessage("labelValue");
     longGauge.getOrCreateTimeSeries(labelValues);
@@ -194,7 +199,8 @@ public class LongGaugeImplTest {
     List<LabelKey> labelKeys =
         Arrays.asList(LabelKey.create("key1", "desc"), LabelKey.create("key2", "desc"));
     LongGaugeImpl longGauge =
-        new LongGaugeImpl(METRIC_NAME, METRIC_DESCRIPTION, METRIC_UNIT, labelKeys);
+        new LongGaugeImpl(
+            METRIC_NAME, METRIC_DESCRIPTION, METRIC_UNIT, labelKeys, EMPTY_CONSTANT_LABELS);
     LongPoint defaultPoint = longGauge.getDefaultTimeSeries();
     defaultPoint.set(-230);
 
@@ -204,6 +210,57 @@ public class LongGaugeImplTest {
     assertThat(metric.getTimeSeriesList().get(0).getLabelValues().size()).isEqualTo(2);
     assertThat(metric.getTimeSeriesList().get(0).getLabelValues().get(0)).isEqualTo(UNSET_VALUE);
     assertThat(metric.getTimeSeriesList().get(0).getLabelValues().get(1)).isEqualTo(UNSET_VALUE);
+  }
+
+  @Test
+  public void withConstantLabels() {
+    List<LabelKey> labelKeys =
+        Arrays.asList(LabelKey.create("key1", "desc"), LabelKey.create("key2", "desc"));
+    List<LabelValue> labelValues =
+        Arrays.asList(LabelValue.create("value1"), LabelValue.create("value2"));
+    LabelKey constantKey = LabelKey.create("constant_key", "desc");
+    LabelValue constantValue = LabelValue.create("constant_value");
+    Map<LabelKey, LabelValue> constantLabels =
+        Collections.<LabelKey, LabelValue>singletonMap(constantKey, constantValue);
+    LongGaugeImpl longGauge =
+        new LongGaugeImpl(METRIC_NAME, METRIC_DESCRIPTION, METRIC_UNIT, labelKeys, constantLabels);
+
+    LongPoint longPoint = longGauge.getOrCreateTimeSeries(labelValues);
+    longPoint.add(1);
+    longPoint.add(2);
+
+    LongPoint defaultPoint = longGauge.getDefaultTimeSeries();
+    defaultPoint.set(100);
+
+    List<LabelKey> allKeys = new ArrayList<>(labelKeys);
+    allKeys.add(constantKey);
+    MetricDescriptor expectedDescriptor =
+        MetricDescriptor.create(
+            METRIC_NAME, METRIC_DESCRIPTION, METRIC_UNIT, Type.GAUGE_INT64, allKeys);
+
+    List<LabelValue> allValues = new ArrayList<>(labelValues);
+    allValues.add(constantValue);
+    List<TimeSeries> expectedTimeSeriesList = new ArrayList<TimeSeries>();
+    TimeSeries defaultTimeSeries =
+        TimeSeries.createWithOnePoint(
+            Arrays.asList(UNSET_VALUE, UNSET_VALUE, constantValue),
+            Point.create(Value.longValue(100), TEST_TIME),
+            null);
+    expectedTimeSeriesList.add(
+        TimeSeries.createWithOnePoint(
+            allValues, Point.create(Value.longValue(3), TEST_TIME), null));
+    expectedTimeSeriesList.add(defaultTimeSeries);
+
+    Metric metric = longGauge.getMetric(testClock);
+    assertThat(metric).isNotNull();
+    assertThat(metric.getMetricDescriptor()).isEqualTo(expectedDescriptor);
+    assertThat(metric.getTimeSeriesList().size()).isEqualTo(2);
+    assertThat(metric.getTimeSeriesList()).containsExactlyElementsIn(expectedTimeSeriesList);
+
+    longGauge.removeTimeSeries(labelValues);
+    Metric metric2 = longGauge.getMetric(testClock);
+    assertThat(metric2).isNotNull();
+    assertThat(metric2.getTimeSeriesList()).containsExactly(defaultTimeSeries);
   }
 
   @Test
@@ -256,7 +313,8 @@ public class LongGaugeImplTest {
         Arrays.asList(LabelValue.create("value1"), LabelValue.create("value2"));
 
     LongGaugeImpl longGauge =
-        new LongGaugeImpl(METRIC_NAME, METRIC_DESCRIPTION, METRIC_UNIT, labelKeys);
+        new LongGaugeImpl(
+            METRIC_NAME, METRIC_DESCRIPTION, METRIC_UNIT, labelKeys, EMPTY_CONSTANT_LABELS);
 
     LongPoint defaultPoint1 = longGauge.getDefaultTimeSeries();
     LongPoint defaultPoint2 = longGauge.getDefaultTimeSeries();