+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 - Add HTTP text format serializer to Tag propagation component.
+- Support constant labels in Gauge APIs.
 
 ## 0.20.0 - 2019-03-28
 - Add OpenCensus Java OC-Agent Trace Exporter.